requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asyncgui==0.6.3
asynckivy==0.6.4
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.2.0
bottle==0.13.2
branca==0.8.1
cachetools==5.5.0
catalogue==2.0.10
cefpython3==66.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
clr_loader==0.2.7.post0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
cymem==2.0.10
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
docutils==0.21.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
faiss-cpu==1.9.0.post1
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
folium==0.19.4
frozenlist==1.5.0
fsspec==2024.12.0
geojson==3.2.0
gitdb==4.0.12
GitPython==3.1.44
googlemaps==4.10.0
greenlet==3.1.1
groq==0.16.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kivy_webview==0.0.0a0
kivymd @ file:///C:/Users/adebi/Desktop/2025_projects/adejoke_chatbot/KivyMD
kvhot==2024.9.12
langcodes==3.5.0
language_data==1.3.0
llama-cloud==0.1.8
llama-index==0.10.0
llama-index-agent-openai==0.1.7
llama-index-cli==0.4.0
llama-index-core==0.10.68.post1
llama-index-embeddings-huggingface==0.5.0
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-huggingface==0.4.2
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.19
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
materialyoucolor==2.0.10
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
narwhals==1.22.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.59.7
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pillow==11.1.0
plyer==2.1.0
preshed==3.0.9
propcache==0.2.1
Protego==0.3.1
protobuf==5.29.3
proxy_tools==0.1.0
psutil==6.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
pypdf==4.3.1
pypiwin32==223
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2024.2
pywebview==5.3.2
pywin32==308
PyYAML==6.0.2
queuelib==1.7.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
Scrapy==2.12.0
sentence-transformers==3.3.1
service-identity==24.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.0
starlette==0.41.3
streamlit==1.41.1
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
text-generation==0.7.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.0
Twisted==24.11.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uv==0.5.18
w3lib==2.2.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
weasel==0.4.1
wrapt==1.17.2
xyzservices==2025.1.0
yarl==1.18.3
zope.interface==7.2
